Split widget: collection row
